+++ src/ServiceProvider.php
@@ -7,6 +7,7 @@ use Statamic\Providers\AddonServiceProvider;
 
 class ServiceProvider extends AddonServiceProvider
 {
+    /** @var array<mixed> */
     protected $tags = [
         BladeComponent::class,
     ];