+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 
-All notable changes to `statamic-blade-components` will be documented in this file
+All notable changes to `statamic-blade-components` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2021-03-26
+
+### Added
+- Render Antler's inside the component tag default slot
+- Added `{{ component:slot }}` to the component tag
+
+### Changed
+- Moved from Statamic `v3.0@beta` to `v3.1`
+
+## 0.0.1 - 2020-08-26
 
 - initial release